src/backend/app/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.12.2
drf-renderer-xlsx==0.4.0
ecdsa==0.14.1
enum-compat==0.0.3
et-xmlfile==1.0.1
future==0.18.2
gunicorn==20.0.4
idna==2.10
intuit-oauth==1.2.4
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysqlclient==2.0.3
oauthlib==3.1.0
openapi-codec==1.3.2
openpyxl==3.0.7
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
python-jose==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
uWSGI==2.0.19.1
whitenoise==5.2.0
